+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0.1003
 - Add Docakable to modify MTP Device Properties (disabled by default)
+- Add support for Bulb with Serial Port, Serial Relay, and Telescope Snap Port
 - Add support for HostMode
 - Enforce the scan of some capabilities for D3200: Liveview, Focusing, ...
 - Fix issues when cammera is abruptly diconnected